requirements.txt
python-dateutil~=2.8.1
pandas~=1.2.3
pytest~=6.2.2